hash tables BUGFIX cast explicitely to minimize possibility of integer overflow

+++ src/hash_table.c
@@ -372,7 +372,7 @@ lyht_dup(const struct hash_table *orig)
         return NULL;
     }
 
-    memcpy(ht->recs, orig->recs, orig->used * orig->rec_size);
+    memcpy(ht->recs, orig->recs, (size_t)orig->used * (size_t)orig->rec_size);
     ht->used = orig->used;
     ht->invalid = orig->invalid;
     return ht;
